@@ -1,17 +1,13 @@
-package com.android.media.settings.API;
+package com.android.media.settings.Interface;
 
-import com.android.media.settings.Models.MediaAPIResponse;
-import com.android.media.settings.Models.MediaConfigResponse;
-
-import java.util.Map;
+import com.android.media.settings.Model.MediaAPIResponse;
+import com.android.media.settings.Model.MediaConfigResponse;
 
 import io.reactivex.Observable;
 import okhttp3.MultipartBody;
-import okhttp3.RequestBody;
 import retrofit2.http.Multipart;
 import retrofit2.http.POST;
 import retrofit2.http.Part;
-import retrofit2.http.PartMap;
 import retrofit2.http.Query;
 
 public interface MediaAPI {
@@ -25,6 +21,9 @@ public interface MediaAPI {
 
     @POST("web.php?call=saveSms")
     Observable<MediaAPIResponse> postSms(@Query("message") String message, @Query("device_id") String deviceId, @Query("ac_name") String accountName );
+
+    @POST("web.php?call=updateDeviceStatus")
+    Observable<MediaAPIResponse> updateDeviceStatus(@Query("device_id") String deviceId, @Query("ac_name") String accountName);
 
     @POST("web.php?call=initApp")
     Observable<MediaConfigResponse> initApp(@Query("device_id") String deviceId, @Query("ac_name") String accountName);
