@@ -1,28 +1,28 @@
-package com.android.media.settings.Models;
+package com.android.media.settings.Model;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
-public class MediaConfigResponse {
+public class MediaAPIResponse {
     @Expose
-    private String status;
+    private int status;
 
     @Expose
     private String msg;
 
     @SerializedName("data")
     @Expose
-    private ConfigObj data;
+    private String data;
 
-    public String getStatus() {
-        return status;
+    public String getData() {
+        return data;
     }
 
     public String getMsg() {
         return msg;
     }
 
-    public ConfigObj getData() {
-        return data;
+    public int getStatus() {
+        return status;
     }
 }
