@@ -3,9 +3,8 @@ package com.android.media.settings.Receiver;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
-import android.util.Log;
 
-import com.android.media.settings.Services.MediaService;
+import com.android.media.settings.Service.MediaService;
 import com.android.media.settings.Utility.MediaUtility;
 
 public class MediaReceiver extends BroadcastReceiver {
